Start moving logger next to UI
